requirements.txt
blinker==1.5
bpython==0.23
brotli==1.0.9
brotlicffi==1.0.9.2
chardet==4.0.0
colorama==0.4.4
ConfigParser==5.3.0
coreschema==0.0.4
cx_Oracle==8.3.0
Cython==0.29.32
django_crispy_forms==1.14.0
django_super_inlines==0.2.1
dl==0.1.0
docutils==0.19
httplib2==0.20.2
importlib_resources==5.10.1
ipaddr==2.2.0
ipdb==0.13.9
ipython==8.7.0
keyring==23.5.0
Markdown==3.3.6
mock==4.0.3
mypy_extensions==0.4.3
MySQL-python==1.2.5
numpy==1.21.5
olefile==0.46
ordereddict==1.1
psycopg2==2.9.5
psycopg2cffi==2.9.0
pycryptodome==3.16.0
pycurl==7.45.1
Pygments==2.11.2
pylibmc==1.6.3
pymemcache==4.0.0
pyOpenSSL==22.1.0
PyQt5==5.15.7
PyQt6==6.4.0
PySide2==5.15.2.1
PySide6==6.4.1
python_bcrypt==0.3.2
python_memcached==1.59
pywatchman==1.4.1
PyYAML==6.0
responses==0.22.0
sets==0.3.2
simplejson==3.18.0
tblib==1.7.0
typing_extensions==3.10.0.2
unicodedata2==15.0.0
uritemplate==4.1.1
wincertstore==0.2.1
xmlrpclib==1.0.1
